move file repos into subdirectory
